Added the ability to cache all HTTP-Content
This is useful because some lecturers tend to restrict access to their
homepages to the university's network. Using this feature, we can
circumvent this. (Should only be used together with the
registration-restriction feature to prevent complaints [^.^]')

Actually, right now, this should not be used at all: I'm still uncertain
how to prevent XSS-attacks originating from the cached contents.

CONSIDER THIS EXPERIMENTAL! I MIGHT DECIDE TO REMOVE THIS FEATURE
AGAIN..