Only deactivate removed sheets instead of removing them
This allows us to handle cases in which editors temporarily remove a
link to a sheet from their homepage and put it back afterwards. Until
now, such a sheet would have automatically been inserted as a new one.
We now recognize it's in reality a sheet we already know and don't
inform users of it's reappearance.

For this we utilize the "timestamp" field of the data table: We set it
to NULL when a sheet disappears (which is consistent as this field's
content means "when was a link found or last updated". Since it's gone
it's rationale to update the column to the value "never")