Fixed an error message, added ZIP-archive caching support (for PHP-entry mode)
This introduces the new function cache_zip_file_contents($url), which
returns an array mapping
	[file base name] => "<Absolute cached url> <File base name>"
for each file in the zip archive.

The code accounts for arising duplicates (p.e. if you have a/foo.zip
and b/foo.zip in the archive) by prepending the file names by a unique
number if necessary. (Though I hope this won't be needed in most cases)

The main reason I added this is because my university's numerics
department uses a software called DocumentTree, which is terribly slow
in serving HTTP requests but comes with a bulk-export function.