Improved caching functionality
The cache now has it's own database table in addition to the already
existing cache directory. Create it using

CREATE TABLE cache (id VARCHAR(40) PRIMARY KEY,
 created_timestamp INT, max_age INT, filename VARCHAR(255))

or just use the updated installation script.

The main improvement of this commit is that we are now able to store
explicit expiration times together with cached data. While this is not
important for the cached output of image.php / combine.php, which can
recreate their output if required, it is of high importance for
PHP-code-based feed definitions. They might create cached contents and
link to it in their output. Now if the cache was flushed, those feeds
would stop working immediately.

Expiration information can be given as a parameter to cache_file(). The
function's signature has changed! Defaults are:
 * Half a year for calls to cache_file() and cache_contents()
 * A day for image.php / combine.php

The cache is only flushed if required, i.e. whenever the cumulated file
sizes exceed a new setting's value (in config.php) the oldest and
already expired contents are removed.