Support for LDAP authentication and account management, refactor newsletter to mail
The small change is that I refactored the confusing
 $user->newsletter
to
 $user->mail
The system is backwards-compatible and automatically updates the users
in the database.

The big one is LDAP-Support. At spline we now have a single-sign-on via
LDAP. I did not want to force users to migrate to LDAP and I did not
want to force new users to embrace it, so my approach is as follows:
 * Local users have precedence over LDAP users
 * New users can not register with names already taken in LDAP, unless
   their password matches the LDAP password. If their desired login
   is present in LDAP, the account is tied to the LDAP.
 * Accounts can be migrated to LDAP on a user's request. For this to
   work, they have to own the user with the matching name in LDAP. They
   have to confirm this by entering the LDAP user's password.
 * Once an account is tied to LDAP, mail and password are no longer
   stored locally, but loaded from the directory server.
 * LDAP authentication works through a simple bind
 * LDAP objects are expected to have uid and mail attributes